requirement.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
cloudinary==1.34.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.3
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
outcome==1.2.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
pyotp==2.9.0
PySocks==1.7.1
pytz==2023.3
razorpay==1.4.1
redis==5.0.0
requests==2.31.0
selenium==4.12.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.10.4
Twisted==23.8.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==1.26.16
websocket-client==1.6.2
wsproto==1.2.0
zope.interface==6.0